requirements.txt
opencv-python
numpy
matplotlib<3.9
pillow-heif
flask
Flask-SQLAlchemy
psycopg2-binary
tqdm
largestinteriorrectangle
tensorflow==2.12.0
deeplabcut[tf]
keras==2.12.0
typing-extensions>=4.6
torch==2.3.0
scipy
huey
redis